Fill _leading buffer before checking for any matches.
Push lines onto the _leading array until it is full
(up to $window_size).  Then shift off one line at a time
and check for matching w/ the email name or qid.

This leaves a problem for cases where the input size
is smaller than the window size: the lines will not be
properly checked.  See the note added to $window_size
setting.  This should only be a problem in testing,
when you use small data files.  Remember to adjust the
window size and keep it smaller than the data file size.

Reformatted code - changed indentation for many lines.