Remove @lines from the grep clause-a logic error.
Use of @lines in the grep clause was a logic error: it will
always find its own element in the array.  Went back to grepping
the entire list of seen qid's when trying to match a qid in a
new line.

Renamed get_seen_lines() to get_seen_qids(), because it returns
hash keys, which are qid's.

Found problem related to window size and not matching lines with
'from=<>'.  Still working on this one.