Improve performance: Remove redundant check of qid.
Only one if() clause is needed in the read loop to check to find
a matching qid, so this second if() clause is not needed.

Remove @lines from the read loop.
I had put it in place of the %_seen hash (in the grep for $match_qid)
& saw that it was a mistake.
It was not used elsewhere, so remove it.

Performance improved greatly after these changes:
time to process the 9999-line i/p file went from 7 sec to 3 sec.