Print only at end of program.
Do not print a group of matching qid's when they are found: save
them in %_seen, and print that hash at end of program.

Check entire buffer for qid matches whenever a line w/ the matching
email addr is found.  Do this check in both the main loop and
in drain_queue().

TBD: Fix problem of duplicate lines being printed.