Prevent duplicate lines that match desired email addr.
Partial fix for duplicate lines in o/p.

In drain_queue, when checking buffer,
do not save any line that matches the line w/ the desired email
addr, because that was already saved in the prior code.

Some duplicate lines still exist, and they are copies of lines
w/ qid's that match the desired qid.

Change code to use @lines_to_drain in a second loop,
instead of rebuilding that array again, which was redundant
in drain_queue().