drain_queue(): find match lines in buffer w/ regex patterns.
Experimenting w/ drain_queue() to fix bugs.
Created other temporary copies of the sub in the file for these
experiments.

Use correct syntax for grep, to match the pattern, to find all lines
with the email addr.  Specify a pattern in the block, by delimiting
it w/ slashes:
  my @lines_with_addr = grep { /$self->{match}/ } @lines_to_drain;

Use a grep stmt to find all lines in the buffer w/ the desired
email addr.

Use a second set of regex pattern matches to find lines in the buffer
w/ a matching qid.

Note: Found that the data files contain some duplicate lines,
w/ same email addrs & qid's.  My code does not create those
duplicate lines; and it should show any duplicate lines that exist
in the data file, when they exist for the desired email addr.