Save line to trailing array if added to %_seen hash.
A line w/ the desired email addr is saved to the %_seen hash.
Copy this line to the @_trailing array also, to ensure that
any matching data in the buffer is found and saved when it is
flushed by drain_queue().

This change causes duplicate lines in the o/p, which must be fixed.