Prevent duplicate lines w/ same email addr & qid.
If some lines have the same email addr & qid, and that is the addr
that is being sought, do not add such lines to the o/p when checking
the buffer for lines whose qid matches a saved qid.  That approach
will cause a duplicate line to be added to the o/p.
Skip that line during the qid check.
Only add it to the o/p when it is shifted off the buffer, and its
email addr matches the reference value.
This design prevents some, but not all, duplicate lines in the o/p
being added by my code.