Prevent duplicate lines from main loop.
Check in main loop for lines w/ matching qid's that have
already been saved, and do not save them again.

Using a global var @emitted, to store the numbers of those lines.
That var is used in main loop, and in drain_queue(), to ensure
that lines saved in the main loop are not duplicated if they
are found by drain_queue().